YANG INFINITO CON SYSTEM SEARCH PRIVATESHOP

+++ Cliente/pack/root/localeinfo.py
@@ -212,7 +212,7 @@ def CutMoneyString(sourceText, startIndex, endIndex, insertingText, backText):
 	if not text:
 		return backText
 
-	if int(text) <= 0:
+	if long(text) <= 0:
 		return backText
 
 	text = str(int(text))
