+++ Source Client/Client/UserInterface/PythonExchange.cpp
@@ -77,12 +77,12 @@ DWORD CPythonExchange::GetChequeFromSelf()
 }
 #endif
 
-void CPythonExchange::SetElkToTarget(DWORD	elk)
+void CPythonExchange::SetElkToTarget(long long elk)
 {
 	m_victim.elk = elk;
 }
 
-void CPythonExchange::SetElkToSelf(DWORD elk)
+void CPythonExchange::SetElkToSelf(long long elk)
 {
 	m_self.elk = elk;
 }
@@ -99,12 +99,12 @@ void CPythonExchange::SetChequeToSelf(DWORD cheque)
 }
 #endif 
 
-DWORD CPythonExchange::GetElkFromTarget()
+long long CPythonExchange::GetElkFromTarget()
 {
 	return m_victim.elk;
 }
 
-DWORD CPythonExchange::GetElkFromSelf()
+long long CPythonExchange::GetElkFromSelf()
 {
 	return m_self.elk;
 }
