+++ Source Client/Client/UserInterface/PythonExchange.h
@@ -30,7 +30,7 @@ class CPythonExchange : public CSingleton<CPythonExchange>
 			TPlayerItemAttribute	item_attr[EXCHANGE_ITEM_MAX_NUM][ITEM_ATTRIBUTE_SLOT_MAX_NUM];
 
 			BYTE					accept;
-			DWORD					elk;
+			long long					elk;
 #ifdef ENABLE_CHEQUE_SYSTEM
 			DWORD					cheque;
 #endif 
@@ -70,14 +70,14 @@ class CPythonExchange : public CSingleton<CPythonExchange>
 		DWORD			GetRaceFromSelf();
 		DWORD			GetRaceFromTarget();
 #endif
-		void			SetElkToTarget(DWORD elk);
-		void			SetElkToSelf(DWORD elk);
+		void			SetElkToTarget(long long elk);
+		void			SetElkToSelf(long long elk);
 #ifdef ENABLE_CHEQUE_SYSTEM
 		void			SetChequeToTarget(DWORD cheque);
 		void			SetChequeToSelf(DWORD cheque);
 #endif 
-		DWORD			GetElkFromTarget();
-		DWORD			GetElkFromSelf();
+		long long		GetElkFromTarget();
+		long long		GetElkFromSelf();
 #ifdef ENABLE_CHEQUE_SYSTEM
 		DWORD			GetChequeFromTarget();
 		DWORD			GetChequeFromSelf();
