+++ Source Client/Client/UserInterface/PythonPlayer.h
@@ -119,11 +119,11 @@ class CPythonPlayer : public CSingleton<CPythonPlayer>, public IAbstractPlayer
 #endif
 			TQuickSlot			aQuickSlot[QUICKSLOT_MAX_NUM];
 			TSkillInstance		aSkill[SKILL_MAX_NUM];
-			long				m_alPoint[POINT_MAX_NUM];
+			long long				m_alPoint[POINT_MAX_NUM];
 			long				lQuickPageIndex;
 
-			void SetPoint(UINT ePoint, long lPoint);
-			long GetPoint(UINT ePoint);
+			void SetPoint(UINT ePoint, long long lPoint);
+			long long GetPoint(UINT ePoint);
 		} TPlayerStatus;
 
 		typedef struct SPartyMemberInfo
@@ -340,8 +340,8 @@ class CPythonPlayer : public CSingleton<CPythonPlayer>, public IAbstractPlayer
 #endif
 
 		void	SetWeaponPower(DWORD dwMinPower, DWORD dwMaxPower, DWORD dwMinMagicPower, DWORD dwMaxMagicPower, DWORD dwAddPower);
-		void	SetStatus(DWORD dwType, long lValue);
-		int		GetStatus(DWORD dwType);
+		void	SetStatus(DWORD dwType, long long lValue);
+		long long	GetStatus(DWORD dwType);
 
 
 		// Item
