+++ sourgame/Server/db/src/MoneyLog.h
@@ -11,10 +11,10 @@ class CMoneyLog : public singleton<CMoneyLog>
 	virtual ~CMoneyLog();
 
 	void Save();
-	void AddLog(BYTE bType, DWORD dwVnum, int iGold);
+	void AddLog(BYTE bType, DWORD dwVnum, long long iGold);
 
     private:
-	std::map<DWORD, int> m_MoneyLogContainer[MONEY_LOG_TYPE_MAX_NUM];
+	std::map<DWORD, long long> m_MoneyLogContainer[MONEY_LOG_TYPE_MAX_NUM];
 };
 
 #endif
