+++ sourgame/Server/game/src/item.cpp
@@ -1489,7 +1489,7 @@ void CItem::SetSocket(int i, long v, bool bLog)
 	}
 }
 
-int CItem::GetGold()
+long long CItem::GetGold()
 {
 	if (IS_SET(GetFlag(), ITEM_FLAG_COUNT_PER_1GOLD))
 	{
@@ -1502,7 +1502,7 @@ int CItem::GetGold()
 		return GetProto()->dwGold;
 }
 
-int CItem::GetShopBuyPrice()
+long long CItem::GetShopBuyPrice()
 {
 	return GetProto()->dwShopBuyPrice;
 }
